Update for building with go 1.16
Due to changes in the module system, go 1.16 no longer tolerates including
local folders via
	import "../system"
Now, on is expected to reference the name of the parent project
	import "xe-guest-utilities/system"

Additional go now expects a go.mod file with listing of external dependencies
and a go.sum with hashes of those.

And go no longer allows '.' within paths that aren't URLs.